Reforma en historias de consultas y mascotas, creacion de opcion de administrador

+++ frontend/src/app/components/shared/navbar/navbar.component.ts
@@ -15,9 +15,12 @@ export class NavbarComponent implements OnInit {
   }
 
   salir(){
-    this.clivet.usuario = [];
-    this.clivet.id_user = "";
-    localStorage.clear();
-    this.router.navigate(['/sesion']);
+    if (confirm('¿Desea cerrar sesión?')) {
+      this.clivet.id_user = "";
+      localStorage.clear();
+      this.router.navigate(['/sesion']);
+    }else{
+      return;
+    }
   }
 }
